requirements.txt
streamlit
Pillow
laspy